requirements.txt
fastapi
uvicorn[standard]
python-dotenv
psutil
PyYAML
qrcode[pil]
